chore: drink 수량을 float으로 변경
- 잔에서 병 단위로 변경(0.5 단위)

+++ src/main/java/com/alc/diary/application/calender/dto/response/FindCalenderDetailResponse.java
@@ -13,14 +13,14 @@ public record FindCalenderDetailResponse(
         List<DrinkModel> drinkModels,
         List<String> images,
         String drinkCondition,
-        int totalDrinkCount
+        float totalDrinkCount
 ) {
 
     public static FindCalenderDetailResponse of(String title, String contents, LocalDateTime drinkStartDateTime, LocalDateTime drinkEndDateTime,
                                              List<DrinkModel> drinkModels, List<String> images, String drinkCondition) {
         return new FindCalenderDetailResponse(
-                title, contents, drinkStartDateTime, drinkEndDateTime, drinkModels, images, drinkCondition, drinkModels.stream()
-                .mapToInt(DrinkModel::getQuantity).sum()
+                title, contents, drinkStartDateTime, drinkEndDateTime, drinkModels, images, drinkCondition, (float) drinkModels.stream()
+                .mapToDouble(DrinkModel::getQuantity).sum()
         );
     }
 }
